reworking architecture of code
