+++ browser/main/Detail/MarkdownNoteDetail.js
@@ -29,10 +29,7 @@ class MarkdownNoteDetail extends React.Component {
         content: ''
       }, props.note),
       isLockButtonShown: false,
-      isLocked: false,
-      fullScreen: false,
-      widthOfNoteDetail: 0,
-      widthOfMainBody: 0
+      isLocked: false
     }
     this.dispatchTimer = null
 
@@ -197,25 +194,7 @@ class MarkdownNoteDetail extends React.Component {
   }
 
   handleFullScreenButton (e) {
-    this.setState({ fullScreen: !this.state.fullScreen }, () => {
-      const noteDetail = document.querySelector('.NoteDetail')
-      const mainBody = document.querySelector('#main-body')
-      const sliderRight = document.querySelector('#slider-right')
-      const sliderLeft = document.querySelector('#slider-left')
-      if (this.state.fullScreen) {
-        this.state.widthOfNoteDetail = noteDetail.style.left
-        this.state.widthOfMainBody = mainBody.style.left
-        noteDetail.style.left = '0px'
-        mainBody.style.left = '0px'
-        sliderRight.style.display = 'none'
-        sliderLeft.style.display = 'none'
-      } else {
-        noteDetail.style.left = this.state.widthOfNoteDetail
-        mainBody.style.left = this.state.widthOfMainBody
-        sliderRight.style.display = 'block'
-        sliderLeft.style.display = 'block'
-      }
-    })
+    ee.emit('editor:fullscreen')
   }
 
   handleLockButtonMouseDown (e) {
